scripts/requirements.txt
psycopg2-binary>=2.9
httpx>=0.27
python-dotenv>=1.0
